requirements/apps.txt
Django==1.2.3
Werkzeug==0.6
BeautifulSoup==3.2.0
django-piston==0.2.2
mimeparse==0.1.3
rdflib==3.0.0
docutils==0.7
pygments==1.3.1
Cython==0.13
lxml==2.2.8
-e git://github.com/django-extensions/django-extensions.git#egg=django-command-extensions
-e git://github.com/robhudson/django-debug-toolbar.git#egg=django-debug-toolbar
